+++ src/ext/java/org/opentripplanner/ext/transmodelapi/mapping/preferences/BikePreferencesMapper.java
@@ -6,14 +6,34 @@ import org.opentripplanner.routing.core.BicycleOptimizeType;
 
 public class BikePreferencesMapper {
 
+  public static final double WALK_BIKE_RELATIVE_RELUCTANCE = 2.7;
+
   public static void mapBikePreferences(
     BikePreferences.Builder bike,
     DataFetcherDecorator callWith
   ) {
     callWith.argument("bikeSpeed", bike::withSpeed);
-    callWith.argument("bikeSwitchTime", bike::withSwitchTime);
-    callWith.argument("bikeSwitchCost", bike::withSwitchCost);
+
+    // These are not supported on the Transmodel API
+    // callWith.argument("bikeSwitchTime", bike::withSwitchTime);
+    // callWith.argument("bikeSwitchCost", bike::withSwitchCost);
+
     callWith.argument("bicycleOptimisationMethod", bike::withOptimizeType);
+
+    // WALK reluctance is used for backwards compatibility, then overridden
+    callWith.argument(
+      "walkReluctance",
+      r -> {
+        bike.withReluctance((double) r);
+        bike.withWalkingReluctance(WALK_BIKE_RELATIVE_RELUCTANCE * (double) r);
+      }
+    );
+
+    // TODO: Override WALK reluctance with BIKE reluctance
+    // callWith.argument("bike.reluctance", r -> {
+    //  bike.withReluctance((double)r);
+    //  bike.withWalkingReluctance(WALK_BIKE_RELATIVE_RELUCTANCE * (double)r );
+    //});
 
     if (bike.optimizeType() == BicycleOptimizeType.TRIANGLE) {
       bike.withOptimizeTriangle(triangle -> {