+++ src/main/java/org/opentripplanner/routing/graph/index/StreetIndex.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
+import javax.annotation.Nullable;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.LineString;
@@ -138,6 +139,7 @@ public class StreetIndex {
     return vertexLinker;
   }
 
+  @Nullable
   public TransitStopVertex findTransitStopVertices(FeedScopedId stopId) {
     return transitStopVertices.get(stopId);
   }