+++ src/main/java/org/opentripplanner/street/model/vertex/TransitStopVertex.java
@@ -3,7 +3,6 @@ package org.opentripplanner.street.model.vertex;
 import java.util.HashSet;
 import java.util.Set;
 import javax.annotation.Nonnull;
-import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.street.model.edge.Edge;
 import org.opentripplanner.street.model.edge.PathwayEdge;
 import org.opentripplanner.transit.model.basic.Accessibility;
@@ -85,5 +84,12 @@ public class TransitStopVertex extends StationElementVertex {
   @Override
   public StationElement getStationElement() {
     return this.stop;
+  }
+
+  /**
+   * Is this vertex already linked to the street network?
+   */
+  public boolean isConnectedToGraph() {
+    return getDegreeOut() + getDegreeIn() > 0;
   }
 }