+++ src/test/java/org/opentripplanner/routing/algorithm/filterchain/DeleteResultHandlerTest.java
@@ -160,7 +160,7 @@ class DeleteResultHandlerTest {
     }
 
     private String tag(Itinerary it) {
-      return it.getSystemNotices().stream().map(n -> n.tag).findFirst().orElse("na");
+      return it.getSystemNotices().stream().map(SystemNotice::tag).findFirst().orElse("na");
     }
   }
 }